+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 - add `providePrismaClientExceptionFilter`, improve exception filter docs
 
+### Changed
+
+- remove deprecated `Prisma.NotFoundError` as it is replaced by `Prisma.PrismaClientKnownRequestError`
+
 # [0.20.0](https://github.com/notiz-dev/nestjs-prisma/compare/v0.19.0...v0.20.0) (2023-01-16)
 
 ### Added